+++ tractor/spawn/__init__.py
@@ -15,12 +15,31 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 '''
-Actor process spawning machinery using
-multiple backends (trio, multiprocessing).
+Actor process spawning machinery using multiple backends.
 
-NOTE: to avoid circular imports, this ``__init__``
-does NOT eagerly import submodules. Use direct
-module paths like ``tractor.spawn._spawn`` or
-``tractor.spawn._entry`` instead.
+Layout
+------
+- `._spawn`: the "core" supervisor machinery — spawn-method
+  registry (`SpawnMethodKey`, `_methods`, `_spawn_method`,
+  `_ctx`, `try_set_start_method`), the `new_proc` dispatcher,
+  and the cross-backend helpers `exhaust_portal`,
+  `cancel_on_completion`, `hard_kill`, `soft_kill`,
+  `proc_waiter`.
+
+Per-backend submodules (each exposes a single `*_proc()`
+coroutine registered in `_spawn._methods`):
+
+- `._trio`: the `trio`-native subprocess backend (default,
+  all platforms), spawns via `trio.lowlevel.open_process()`.
+- `._mp`: the stdlib `multiprocessing` backends —
+  `'mp_spawn'` and `'mp_forkserver'` variants — driven by
+  the `mp.context` bound to `_spawn._ctx`.
+
+Entry-point helpers live in `._entry`/`._mp_fixup_main`/
+`._forkserver_override`.
+
+NOTE: to avoid circular imports, this ``__init__`` does NOT
+eagerly import submodules. Use direct module paths like
+``tractor.spawn._spawn`` or ``tractor.spawn._trio`` instead.
 
 '''
